requirements.txt
fastapi
uvicorn
bcrypt
PyJWT
python-dotenv
loguru
sqlalchemy
pymysql